requirements.txt
openai>=1.30.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.150.0
python-dotenv>=1.0.1
requests>=2.32.0
beautifulsoup4>=4.12.0
Pillow>=10.0.0
